Apply useSWR for project
