+++ src/components/ProjectReturn.js
@@ -1,26 +1,12 @@
-import axios from 'axios'
 import { makeStyles } from '@material-ui/core/styles'
 import Card from '@material-ui/core/Card'
 import CardContent from '@material-ui/core/CardContent'
 import CardMedia from '@material-ui/core/CardMedia'
 import Typography from '@material-ui/core/Typography'
 import Button from '@material-ui/core/Button'
-import { APIEndpoints } from '../constants'
 
-export default function ProjectReturn({ projectReturn, user }) {
+export default function ProjectReturn({ projectReturn, handleSupportButtonClick }) {
   const classes = useStyles()
-
-  const handleSupportButtonClick = async event => {
-    event.preventDefault()
-
-    try {
-      const token = await user.getIdToken()
-      const config = { headers: { authorization: `Token ${token}` } }
-      await axios.post(APIEndpoints.PROJECT_SUPPORTS, { project_return_id: projectReturn.id }, config)
-    } catch (error) {
-      console.log(error)
-    }
-  }
 
   return (
     <div>
@@ -47,7 +33,7 @@ export default function ProjectReturn({ projectReturn, user }) {
             {projectReturn.description}
           </Typography>
 
-          <Button onClick={handleSupportButtonClick} fullWidth variant='contained' color='secondary'>
+          <Button onClick={handleSupportButtonClick(projectReturn.id)} fullWidth variant='contained' color='secondary'>
             支援する
           </Button>
         </CardContent>
